// ahm-lc-vin-bom-client-master/ahm-lc-vin-bom-client-master/lc-vin-bom-client/src/main/js/src/app/models/vin-bom-part-dto.ts
export class VinBomPartDto {
    productSpecCode: string;
    letSystemName: string;
    modelYearCode: string;
    modelCode: string;
    modelTypeCode: string;
    basePartNumber: string;
    dcPartNumber: string;
    description: string;
    effectiveBeginDate: string;
    effectiveEndDate: string;
    active:string;
}